Converts to 4-spaces. Cleans up HTML/CSS.
This is a pretty major code thrash change. Changes everything to
4-spaces, runs CSSComb on all our CSS, adds newlines at ends of files,
and fixes some general CSS rules.

Be very careful when merging that you don’t lose anything. I’m happy to
do a pair merge with anyone who wants it when merging, as this is a
seriously hairy merge. It’s too bad GitHub doesn’t do whitespace-aware
diffs.