ChooseEnergyTheme CSS Bugs fix
1 item of note: I added another piece of html in for those pages with
large hero images. This is optional but I think it makes things look
nicer, and is a step towards the parralax-esque hero images that we’ve
always wanted. I think the hero images are as good as they’ll
realistically get without incorporating some js or a real look at
responsiveness.